+++ sql-psi/src/test/kotlin/com/alecstrong/sql/psi/test/fixtures/TestHeadlessParser.kt
@@ -58,7 +58,7 @@ object TestHeadlessParser {
 private object TestLanguage : Language("Test")
 
 private object TestFileType : LanguageFileType(TestLanguage) {
-  override fun getIcon() = AllIcons.Icon
+  override fun getIcon() = AllIcons.FileTypes.Text
 
   override fun getName() = "Test File"